refactor: recommend 로직 분리
